relationshipFixes
Fixed the relationship methods. We were having trouble with ID. I talked to Dr. Schwartz, she said that we can have the user give the ID. I also researched a way to use maps to hold the relationships by the Integer type. This holds the integer in an object allowing it to be able to be a key. THIS NEEDS TO BE TESTED.